argparse.js fails 2 help related tests in test_argparse.py
TestParserDefault42 sets a parser argumentDefault
The error is that the namespace includes {help: 42, version: 42}
Correction: add defaultValue: $$.SUPPRESS when defining --help and --version
in ArgumentParser().

TestOptionalsAlternatePrefixCharsAddedHelp sets prefixChars.
If prefixChars does not include '-', setting ['-h','--help'] gives an error.
Correction: set a defaultPrefix that is either '-' or the first prefixChar,
and use this instead of '-h'