style(tests): Standardized format via eslint.
As a side note, for some reason eslint was trying to tell me
that I had lines longer than 120 characters when my preffered line
length is set to 80 in my editor and I usually never exceede that.

Apparently after some fiddling around with it, it seems the issue
was with my indentation format: I use hard tabs by default.
I'll probably file a bug report with ESLint to verify / fix, but
just in case I forget I'll leave this in the commit history.

Beware the crusty `--fix`, it'll break yer code m8ies.